backend/requirements.txt
Flask
Flask-CORS
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
deepgram-sdk
requests
pymilvus==2.4.3
sentence-transformers
transformers
torch
PyPDF2
ai71